+++ include/sf33rd/Source/Compress/zlibApp.h
@@ -4,6 +4,6 @@
 #include "types.h"
 
 void zlib_Initialize(void *tempAdrs, s32 tempSize);
-s32 zlib_Decompress(void *srcBuff, s32 srcSize, void *dstBuff, s32 dstSize);
+ssize_t zlib_Decompress(void *srcBuff, s32 srcSize, void *dstBuff, s32 dstSize);
 
 #endif
