+++ include/sf33rd/Source/PS2/mc/knjsub.h
@@ -3,7 +3,7 @@
 
 #include "types.h"
 
-void KnjInit(u32 type, u32 adrs, u32 disp_max, u32 top_dbp);
+void KnjInit(u32 type, uintptr_t adrs, u32 disp_max, u32 top_dbp);
 void KnjFinish();
 void KnjSetSize(s32 dispw, s32 disph);
 void KnjLocate(s32 x, s32 y);
