+++ src/anniversary/sf33rd/AcrSDK/ps2/flps2ps.c
@@ -15,6 +15,6 @@ s32 flPS2psAddQueue(QWORD *dma_data) {
     dma_data->I32[2] = 0x13000000;
     dma_data->I32[3] = (qwc - 1) | 0x51000000;
     flPS2_Mem_move16_16A(dma_data, (void *)dst, qwc);
-    flPS2DmaAddQueue2(0, ((u32)dst & 0x0FFFFFFF) | 0x40000000, (u32)dst, &flPs2VIF1Control);
+    flPS2DmaAddQueue2(0, ((uintptr_t)dst & 0x0FFFFFFF) | 0x40000000, (uintptr_t)dst, &flPs2VIF1Control);
     return 1;
 }
